+++ packages/jtcsv-codemod/transforms/index.js
@@ -1,3 +1,4 @@
 module.exports = {
   'papaparse-to-jtcsv': require('./papaparse-to-jtcsv'),
+  'csvtojson-to-jtcsv': require('./csvtojson-to-jtcsv'),
 };